Create parsing.py, reduce file reads
Each file being modified is now read at most once.
Code has been moved from fixes.py to parsing.py, which is intended to
hold logic for parsing and understanding code.
The resulting formatting is overwritten to the original file only if it
is valid python, regardless of whether it originally was or not.