Add initial rmspace run
This solves a bug where trailing whitespace on the final line would make
a regex unable to find a line for that character. Formatting with
rmspace solves this.
All files should end with a newline according to unix definitions, so I
don't think I should have to preserve any exceptions to this.