Drop support for python3.8, bump version to 6
Python support is dropped pre-python3.8. This is because there are
significant differences in the ast strucutures from 3.8 to 3.9, and
anyone who is dabbling with automated python refactoring is likely also
able to install a python interpreter running 3.9.

In some sense, this may be thought of as technical debt to be solved in
the future, but it is very low priority stuff. The purpose of pyrefact
is to be bleeding-edge, and python3.8 is just not.