Fix bug in is_blocking
Unfortunately we can seldom tell that a for loop is blocking. We could
imagine parsing the iterator with eval, but this risks arbitrary code
execution and I think you may want to be able to run pyrefact on
untrusted code. So, if we can't show that the iterator is empty, we also
can't say whether any exceptions or similar inside make it blocking.