Alter preference of contains on inlined comps
When finding inlined set comprehensions, pyrefact will prefer:
Set comprehensions if the expression can be parsed with ast.literal_eval
Generator comprehensions otherwise
The intent of this is that expressions that the python interpreter can
trust not to have a side effect will be evaluated only once at compile
time, but other expressions will be evaluated once every time they're
used. So, if evaluated once at compile time it makes sense to make them
sets, but otherwise it is better to make them generators to avoid some
hashing.