Simplify subscript comprehensions
Comprehensions over subscripts are replaced by regular loops. For
example, [a[i] for i in range(len(a))] is replaced with list(a). This
enables the matrix multiplication replacements to become a lot more
powerful.
In addition, a feature to simplify matrix transposes is added, that
currently just replaces transposes on the form A.T.T with A.T, i.e. it
only replaces directly chained transposes.