Simplify matrix operations, more tests.
Pyrefact will now simplify matrix multiplications where everything is
transposed, i.e. (A.T * B.T).T will become B * A. This is because the
logic that replaces comprehensions with matmuls tends to often create
this pattern, and it doesn't look very nice.
Tests are added for this, and the tests for replacing comprehensions
with matmul calls are extended.