Remove test cases that auto-add indentation
An integration test case is added that asserts that a loop performing a
dot product by adding to a constant can be converted all the way to a
np.dot() call.
import numpy as np is deliberately left out, and the test asserts that
pyrefact can add it back.
This highlights, however, that overindented code shouldn't be processed
the same as regular code. Since overindented code is likely a part of a
function or similar, it doesn't make sense to add imports and rename
variables etc there. But then the test cases for over-indented code are
not as straightforward either, and for now I simply delete them.