Refactor test folder structure and logic
Tests are grouped in two folders; unit and integration. All unit or
integration test files are renamed to test_xyz, where xyz is the
function that they test. Some logic is rewritten in tests/main.py to
work with this, and the sys.path modifications that are made in all the
test files are rewritten so that the tests can still be run directly as
well.