Fixes for replace_with_filter
Looks like I made a couple mistakes in implementing this. The opposite
branch was not properly checked, which led to it wrongly ignoring
possible code that should have run if the filter callable condition was
True.
It now also supports replacing to filter(None, iterable), for the common
case of an immediate "if x:" following "for x in iterable:"