Create parsing.match_template
The function match_template is added to parsing. This is a primitive
pattern matcher, and will check if nodes comply with a provided ast
structure that should be very similar to what is provided by ast.dump().
The function is implemented throughout parsing and replaces lots of
calls to isinstance().
replace_implicit_dot() and replace_implicit_matmul() are is also
reimplemented with the function, so that it will have a use outside of
parsing and not be made private by pyrefact itself.