Remove black formatting of entire files
Pyrefact will no longer format the whole input with black, but only
parts of expressions that were generated by unparse(), or lines that
were too long when everything else had completed.

The purpose of this change is so that pyrefact should to the minimum
needed changes, so that the diff for someone looking at a file that is
not already pre-formatted with black on the same settings should not
be polluted by a bunch of formatting changes. Then users have greater
freedom to choose whatever formatting style they like.