Add processing.fix decorator and Range and Rewrite
The processing.fix decorator is added, which wraps a function yielding
(old, new) tuples into replacements, and applies those replacements.
This aims to centralize some of the logic that is common for many of the
fix functions, i.e. that something is replaced with something else. Two
classes Range and Rewrite are also added, to have more well-defined
types when doing a replacement, but I don't find them super useful just
yet, they may go away in the future.