Delete sql-related code
All sql-related code is deleted.

The reason for this decision is that I'm finding it hard to find
something to check whether code is valid sql syntax, and I also don't
think it's feasible for me to write something like this on my own.
If this changes in the future, this can be reverted.