Move creating abstractions after hot loop
Abstractions now get only one chance to be created, and this will be
after the main processing loop inside main(). The reason for this is
that abstractions can have a way of first creating an abstraction for
some code that is then simplified in some other way, leading to an
abstracted function that doesn't do much. A good example of this was
with the new dict simplification functions.