Add move_before_loop
Pyrefact will now move code that does not depend on the content of for
loops before the for loops, reducing duplicate work.
I'm aware that loops may or may not be entered at all, meaning that this
may cause code to execute when it would not previously have. Which is
completely fine as it should be very uncommon and a definite
anti-pattern wherever it exists.