Make move_before_loop much more defensive
The feature is made to need much greater certainty before moving any
code in front of a nearby loop. It will now treat attribute accesses,
subscripts and augmented assigns (i.e. +=) as other assignments of the
same variable, and will therefore not move it.

This is a shame, since it makes the feature much less powerful.