Change all dependencies to >=
Pyrefact now requires a minimum version, instead of an exact version.
These exact versions do not matter particularly much to me, and this is
intended to make it easier to install pyrefact into an environment with
lots of hard dependencies for other things.

astunparse is kept fixed, as I have quite a bit of code written to work
with specific strange quirks of astunparse==1.6.3.