Bump version to 76
simplify_assign_immediate_return is added, which will remove variables
that are defined only once, and immediately before the same variable is
returned. Instead whatever the variable was assigned to will be returned
directly.

In addition, an issue where pyrefact would sometimes change the
indentation of code in an incorrect way is solved.