Implement compile_template() and find_replace()
In an effort to make the pyrefact code more robust and more readable,
two new functions are added to pyrefact, and implemented throughout the
library:
compile_template: This converts python-like code into a template usable
by the pyrefact pattern-matcher.
find_replace: This searches for a compiled or uncompiled (it will then
compile it first) template, and returns an iterator of ranges of code
that matched that template, as well as a formatted replacement given by
the replace argument. This works nicely with @processing.fix.

These two functions are implemented throughout pyrefact. An additional
benefit is that much of the python3.8-specific code has gone away, as
compile_template() will handle whatever version-specific details that
may exist. One can hope that this may mean that future pyrefact versions
will continue to work for longer into the future.