Make fix more conservative
delete_unused_functions_and_classes is made more conservative, to
prevent FN deletions of code that is actually used. If we have typing
inference in the future, this may be reversed.

In addition, the tests are extended somewhat, although they do not test
the changes made here as those only apply when inferring if something is
used across files.