Add custom syntax for zeroorone etc
Special syntax is added to the template parser, so that `+*?` can be
used in a regex-like way. The special wildcard `...` is added, which
matches anything in a non-common way (e.g, if there are multiple matches
they do not need to be identical).

Together, they can be combined in these ways:
{{name}}: A regular wildcard, as before
{{name?}}: A regular wildcard, match 0 or 1 times
{{name*}}: A regular wildcard, match 0 or more times
{{name+}}: A regular wildcard, match 1 or more times
{{...}}: Match anything, exactly 1 times
{{...?}}: Match anything, 0 or 1 times
{{...*}}: Match anything, 0 or more times
{{...+}}: Match anything, 1 or more times